% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivalBoard.R
\name{arrivalBoard}
\alias{arrivalBoard}
\title{Fetch an arrival board}
\usage{
arrivalBoard(stopID = "008011160", date = substr(Sys.time() + 1800, 1, 10),
  time = substr(Sys.time() + 1800, 12, 16), authKey, refs = FALSE)
}
\arguments{
\item{stopID}{character, \code{stopID} found with \code{getStopID}.}

\item{date}{character, set the date to fetch an arrival board for in \code{"YYYY-MM-DD"} format. Default uses \code{Sys.time()}.}

\item{time}{character, set the earliest arrival time in \code{"HH:MM"} format. Default uses \code{Sys.time()}.}

\item{authKey}{character, authentication key for the DB-Open-Data timetable API.}

\item{refs}{logical, if \code{TRUE} additional references will be fetched. Default is \code{FALSE}.}
}
\value{
A data.frame if \code{refs = FALSE} or a list if {refs = TRUE}.
}
\description{
Fetch an arrival board from the DB-Open-Data timetable API and corresponding references.
}
\details{
\code{arrivalBoard()} will fetch an arrival board for a given \code{date} containing the next 20 arrivals at a certain station (\code{stopID}) beginning from the given \code{time}. Additionally further train-specific references can be fetched with \code{refs = TRUE}. These can be used with \code{journeyDetails()}.
}
\author{
Philipp Ottolinger
}

