% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStopID.R
\name{getStopID}
\alias{getStopID}
\title{Get the \code{stopID}}
\usage{
getStopID(station = "Berlin Hbf", authKey)
}
\arguments{
\item{station}{character, search a station by name. Default is \code{"Berlin Hbf"}.}

\item{authKey}{character, authentication key for the DB-Open-Data timetable API.}
}
\value{
A data.frame containing the station name and the respective \code{stopID}.
}
\description{
Get the \code{stopID} needed for the DB-Open-Data timetable API.
}
\details{
Because the timetable API cannot handle station names one has to fetch the respective \code{stopID} first. This \code{stopID} can be used to refer to a certain station.
}
\author{
Philipp Ottolinger
}

