% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haltestellendaten}
\alias{haltestellendaten}
\title{List of all stations of the DB Station&Service AG.}
\format{A \code{data.frame} with 6598 observations and 6 variables:
\describe{
  \item{EVA_NR}{Internal station ID.}
  \item{DS100}{Corresponding railway system.}
  \item{NAME}{Name of the station.}
  \item{VERKEHR}{Type of traffic: "FV" = also long-distance traffic, "RV" = regional traffic only, "nur DPN" = private regional traffic only.}
  \item{LAENGE}{The station's longitude (WGS84).}
  \item{BREITE}{The station's latitude (WGS84).}
}}
\source{
\url{http://download-data.deutschebahn.com/static/datasets/haltestellen/D_Bahnhof_2016_01_alle.csv}
}
\usage{
haltestellendaten
}
\description{
List of all stations of the DB Station&Service AG.
}
\references{
\url{http://data.deutschebahn.com/dataset/data-haltestellen}
}
\keyword{datasets}

