% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeyDetails.R
\name{journeyDetails}
\alias{journeyDetails}
\title{Get details on a specific train}
\usage{
journeyDetails(ref)
}
\arguments{
\item{ref, }{references URL from \code{departureBoard(..., refs = TRUE)$departureBoardRef} or \code{arrivalBoard(..., refs = TRUE)$arrivalBoardRef}.}
}
\description{
Get additional information on a specific train like the rooute, stops, arrival and departure times.
}
\details{
\code{journeyDetails()} lets you get further information on a specific train. It requires a reference URL fetched with \code{departureBoard()} or \code{arrivalBoard()} and fetches the respective journey details from the API.
}
\examples{
\dontrun{
dep <- departureBoard(stopID = "008002549",
                      date = "2016-12-01",
                      time = "08:00",
                      authKey = authKey,
                      refs = TRUE)
journeyDetails(dep$departureBoardRef$Ref[1])
}
}
\author{
Philipp Ottolinger
}

