% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luftschadstoffkataster}
\alias{luftschadstoffkataster}
\title{Air contaminant cadastre (extract).}
\format{A \code{data.frame} with 445437 observations and 16 variables:
\describe{
  \item{id}{Section id.}
  \item{long}{Longitude of the section.}
  \item{lat}{Latitude of the section}
  \item{order}{Order of the sections.}
  \item{hole}{?}
  \item{piece}{?.}
  \item{group}{Group of the sections.}
  \item{LayerName}{?}
  \item{EntityType}{?}
  \item{EntityColo}{?}
  \item{BlockName}{?}
  \item{EntityIsVi}{?}
  \item{TextJustif}{?}
  \item{MapName}{?}
  \item{NOX_kg_a}{Nitric oxide emission, kg per year.}
  \item{PM10V_kg_a}{Particle emission, kg per year.}
}}
\source{
\url{http://download-data.deutschebahn.com/static/datasets/luftqualitaet/Luft_V_Fpl2014.zip}
}
\usage{
luftschadstoffkataster
}
\references{
\url{http://data.deutschebahn.com/dataset/luftschadstoffkataster}
}
\keyword{datasets}

