% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netzradar}
\alias{netzradar}
\title{The German long-distance traffic railway network with information on mobile communications coverage.}
\format{A \code{data.frame} with 119327 observations and 23 variables:
\describe{
  \item{id}{Track section id.},
  \item{lon}{Longitude of the track section.}
  \item{lat}{Latitude of the track section}
  \item{all_stability}{Network stability for all providers.}
  \item{all_measurements}{Number of measurements for all providers.}
  \item{all,no4g_stability}{Network stability for all providers, without LTE.}
  \item{all,no4g_measurements}{Number of measurements for all providers, without LTE.}
  \item{t-mobile_stability}{Network stability for provider T-Mobile.}
  \item{t-mobile_measurements}{Number of measurements for provider T-Mobile.}
  \item{t-mobile,no4g_stability}{Network stability for provider T-Mobile, without LTE.}
  \item{t-mobile,no4g_measurements}{umber of measurements for provider T-Mobile, without LTE.}
  \item{vodafone_stability}{Network stability for provider Vodafone.}
  \item{vodafone_measurements}{Number of measurements for provider Vodafone.}
  \item{vodafone,no4g_stability}{Network stability for provider Vodafone, without LTE.}
  \item{vodafoneno4g_measurements}{umber of measurements for provider Vodafone, without LTE.}
  \item{e-plus_stability}{Network stability for provider E-Plus.}
  \item{e-plus_measurements}{Number of measurements for provider E-Plus.}
  \item{e-plus,no4g_stability}{Network stability for provider E-Plus, without LTE.}
  \item{e-plus,no4g_measurements}{umber of measurements for provider E-Plus, without LTE.}
  \item{o2_stability}{Network stability for provider o2.}
  \item{o2_measurements}{Number of measurements for provider o2.}
  \item{o2,no4g_stability}{Network stability for provider o2, without LTE.}
  \item{o2,no4g_measurements}{umber of measurements for provider o2, without LTE.}
}}
\source{
\url{http://download-data.deutschebahn.com/static/datasets/netzradar/connectivity_2016_11.geojson}
}
\usage{
netzradar
}
\description{
The German long-distance traffic railway network with information on mobile communications coverage.
}
\references{
\url{http://data.deutschebahn.com/dataset/data-netzradar}
}
\keyword{datasets}

