% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stationsdaten}
\alias{stationsdaten}
\title{List of all stations with addresses.}
\format{A \code{data.frame} with 5372 observations and 10 variables:
\describe{
  \item{Bundesland}{The respective federal state.},
  \item{BM}{The respective station management.}
  \item{Bf..Nr.}{Unique station number.}
  \item{Station}{Name of the station.},
  \item{Bf.DS.100.Abk.}{Corresponding railway system.}
  \item{Kat..Vst}{Station category.},
  \item{Straße}{Street, as part of address.}
  \item{PLZ}{Postcode, as part of address.},
  \item{Ort}{City, as part of address.}
  \item{Aufgabenträger}{Constractor.}
}}
\source{
\url{http://download-data.deutschebahn.com/static/datasets/stationsdaten/DBSuS-Uebersicht_Bahnhoefe-Stand2016-07.csv}
}
\usage{
stationsdaten
}
\references{
\url{http://data.deutschebahn.com/dataset/data-stationsdaten}
}
\keyword{datasets}

