% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable_reference.R
\name{timetable_reference}
\alias{timetable_reference}
\title{Get detailed information on a specific train}
\usage{
timetable_reference(reference)
}
\arguments{
\item{reference}{character string. A reference URL from \code{timetable_arrivals()} or \code{timetable_departures()}.}
}
\value{
A \code{list} with journey details.
}
\description{
Use a reference URL from either a departure or an arrival board to get more detailed information on a train.
}
\examples{
\dontrun{
tt <- timetable_arrivals(station = berlin, direction = hamburg,
                         date = "2016-12-31", time = "12:00",
                         authkey = authKey, reference = TRUE)
timetable_reference(as.character(tt$ref[1]))
}
}
\author{
Philipp Ottolinger
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\seealso{
\code{\link{timetable_departures}}, \code{\link{timetable_arrivals}}, \code{\link{timetable_station}}
}

