% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable_station.R
\name{timetable_station}
\alias{timetable_station}
\title{Find a station and its internal ID}
\usage{
timetable_station(pattern, authkey, first = TRUE, lang = "de",
  format = "json")
}
\arguments{
\item{pattern}{character string. Pattern used to find a specific train station, e.g. 'Berlin'.}

\item{authkey}{character string. Authentication key to access the Timetable-API. See Note.}

\item{first}{logical. If TRUE, the ID of the first matching station is returned. If FALSE, a list of stations is returned.}

\item{lang}{character string. Set the API language. Default is 'de'.}

\item{format}{character string. Set the API format. Default is 'json'.}
}
\value{
A character string if \code{first = TRUE}. Otherwise a \code{data.frame}.
}
\description{
Use a search pattern to find a specific train station in Germany and its internal ID. The internal ID can be used to query the DB Timetable-API.
}
\note{
To use this function and to access the DB Timetable-API, you need to have an authentication key. Authentication keys can be requested from \code{DBOpenData@deutschebahn.com}.
}
\examples{
\dontrun{
timetable_station(pattern = "Berlin", authkey = authKey)
timetable_station(pattern = "Hamburg", authkey = authKey, first = FALSE)
}
}
\author{
Philipp Ottolinger
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\seealso{
\code{\link{timetable_departures}}, \code{\link{timetable_arrivals}}, \code{\link{timetable_reference}}
}

