% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeyDetails.R
\name{journeyDetails}
\alias{journeyDetails}
\title{Get details on a specific train}
\usage{
journeyDetails(refs, Train)
}
\arguments{
\item{refs}{data.frame, references table from \code{departureBoard(..., refs = TRUE)$departureBoardRef} or \code{arrivalBoard(..., refs = TRUE)$arrivalBoardRef}.}

\item{Train}{character, train name as found in departure board or arrival board.}
}
\description{
Get additional information on a specific train like the rooute, stops, arrival and departure times.
}
\details{
\code{journeyDetails()} lets you get further information on a specific train. It requires a table with reference URLs fetched with \code{departureBoard()} or \code{arrivalBoard()}. The function will choose the URL corresponding to the given train name (\code{Train}) and fetch the data from the API.
}
\author{
Philipp Ottolinger
}

